saving changes from codesandbox.io. Refactored bookGradShootPage -> BookShootPage; created booking form; updated page description.
