+++ src/App.js
@@ -2,7 +2,7 @@ import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import ContactPage from "./components/ContactPage.js";
-import BookGradShootPage from "./components/BookGradShootPage.js";
+import BookShootPage from "./components/BookShootPage.js";
 import LandingPage from "./components/LandingPage.js";
 import WorksPage from "./components/WorksPage.js";
 import Header from "./components/Header.js";
@@ -16,7 +16,7 @@ function App() {
         <Routes>
           <Route path="/" element={<LandingPage />} />
           <Route path="/contact" element={<ContactPage />} />
-          <Route path="/grad-shoot-booking" element={<BookGradShootPage />} />
+          <Route path="/book-a-shoot" element={<BookShootPage />} />
           <Route path="/works" element={<WorksPage />} />
           <Route path="/works" element={<WorksPage />} />
           <Route
