--- src/components/BookGradShootPage.css
@@ -1,71 +1,0 @@
-.bookGradShootPage {
-    display: flex;
-    flex-direction: column;
-}
-
-.clientInputFieldsContainer {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-}
-
-.fullWidth {
-    width: 500px;
-}
-
-.clientInputFields {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    row-gap: 5px;
-    padding: 0 0 20px 0;
-    /* max-width: 50%; */
-    row-gap: 10px;
-    width: 500px
-}
-
-.selectDropDown {
-    margin: 10px;
-}
-
-.pageTitleContainer {
-    text-align: center;
-}
-
-.selectContactMethodContainer {
-    min-width: none;
-}
-
-.selectContactMethod {
-    
-}
-
-.clientSchool {
-    padding-top: 15px;
-}
-
-.homeButton {
-    max-width: 10px;
-}
-
-.bookShootPage__bookButtonContainer {
-    display: flex;
-}
-
-.bookShootPage__homeButton {
-    align-self: left ;
-}
-
-.contactContainer {
-    display: flex;
-    align-items: flex-end;
-    justify-content: space-between;
-}
-
-.contactInput {
-    align-self: flex-end;
-}
-
-.contactRadioButtonsContainer {
-    /* padding-right: 20px; */
-}
